REFAC: Avoid checkWinner repetitive implementation
In previous iteration each `checkCross` methods created a loop that was
Essentially the same, among all of them, except for the YX coordinates.
In this implementation `checkWinner()` creates a single main loop
Wich the checkCross methods use and receives its current iteration `i`
To return the appropiate XY coordinates for Current Spot and Next Spot
Wich is the only thing that varies among them.
checkWinner's main loopchecks sucesive spots
Looking for a 3-in-line Cross-out